pyproject.toml
[tool.poetry]
name = "slack-bot"
version = "0.1.0"
description = ""
authors = ["Harsh Pratap Singh <harsh@Harshs-MacBook-Pro.local>"]
readme = "README.md"
packages = [{include = "slack_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.1"
aiosignal = "^1.2.0"
async-timeout = "^4.0.2"
attrs = "^22.1.0"
charset-normalizer = "^2.1.1"
frozenlist = "^1.3.1"
idna = "^3.3"
multidict = "^6.0.2"
python-dotenv = "^0.21.0"
slackclient = "^2.9.4"
yarl = "^1.8.1"
Flask = "^2.2.2"
slackeventsapi = "^3.0.1"
requests = "^2.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
